Fix jumping between cells
Let there be air Cells A and B, and solid Cell C.
Where Cell A borders cells B and C on its bottom,
and Cell B borders Cell C on it's right side. If there was an entity
with its top left in Cell A and the rest in Cell B, when it tried
to move left, its position would snap so the bottom lines up with the
bottom on Cell A.

I thought I had fixed that issue, but I guess it was reverted.